* nbody:    * added VR rendering + controller support    * unified_renderer: will now render into an intermediate image/renderbuffer, then blit it to the screen (this is to make sure that blending works properly with wide-gamut/HDR)    * shader: changed point size range to [1.5, 65.0] (from [2.0, 255.0])    * shader: multi-view rendering support for VR (now has two model-view and perspective matrices)    * CMake: added MinGW support + added release mode optimization flags  * updated config with VR settings + enabled wide-gamut/HDR by default

+++ nbody/src/gl_renderer.cpp
@@ -42,7 +42,7 @@ static void create_textures() {
 		array<uint32_t, texture_size.x * texture_size.y> pixel_data;
 		for(uint32_t y = 0; y < texture_size.y; ++y) {
 			for(uint32_t x = 0; x < texture_size.x; ++x) {
-				float2 dir = (float2(x, y) / texture_sizef) * 2.0f - 1.0f;
+				float2 dir = (uint2(x, y).cast<float>() / texture_sizef) * 2.0f - 1.0f;
 #if 1 // smoother, less of a center point
 				float fval = dir.dot();
 #else
